improved start.sh, readme
1. correctly implemented start.sh so that user can change CONTEXT_PATH and SSL environment variables in docker-compose.yml
2. Removed `ln -s` error on the second and subsequent launches.
3. improved readme

+++ start.sh
@@ -1,11 +1,23 @@
 #!/bin/sh
 
+if [ SSL == "no" ]; then
+  HTTPS_PORT=0
+else
+  HTTPS_PORT=4041
+fi
+#echo $CONTEXT_PATH
+#echo $HTTPS_PORT
+
 addgroup -g ${GID} subsonic && adduser -h /subsonic -s /bin/sh -D -G subsonic -u ${UID} subsonic
 
 mkdir -p /config/transcode
 
-ln -s /usr/bin/ffmpeg /config/transcode/ffmpeg
-ln -s /usr/bin/lame /config/transcode/lame
+if [ ! -e /config/transcode/ffmpeg ]; then
+  ln -s /usr/bin/ffmpeg /config/transcode/ffmpeg
+fi
+if [ ! -e /config/transcode/lame ]; then
+  ln -s /usr/bin/lame /config/transcode/lame
+fi
 
 chown -R subsonic:subsonic /config
 
@@ -19,12 +31,11 @@ java -Xmx${JVM_MEMORY}m \
   -Dsubsonic.home=/config \
   -Dsubsonic.host=0.0.0.0 \
   -Dsubsonic.port=4040 \
-  -Dsubsonic.httpsPort=4041 \
-  -Dsubsonic.contextPath=/ \
+  -Dsubsonic.httpsPort=$HTTPS_PORT \
   -Dsubsonic.defaultMusicFolder=/media \
   -Dsubsonic.defaultPodcastFolder=/podcasts \
   -Dsubsonic.defaultPlaylistFolder=/playlists \
-  -Dsubsonic.contextPath=/ \
+  -Dsubsonic.contextPath=$CONTEXT_PATH \
   -Djava.awt.headless=true \
   -verbose:gc \
   -jar subsonic-booter-jar-with-dependencies.jar
